requirements.txt
openai
requests
beautifulsoup4
python-slugify
python-dotenv
